requirements.txt
cycler==0.10.0
h5py==3.2.1
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
numba~=0.54.0